# swagger.yaml
openapi: 3.0.0
info:
  title: social app
  description: First social application - Deepz
  version: 0.1.1

servers:
  - url: "{protocol}://localhost:5000/api/"
    description: localhost version of the app
    variables:
      protocol:
        enum:
          - http
          - https
        default: http    

components:
  securitySchemes:
    BasicAuth:
      type: http
      scheme: basic
    BearerAuth:
      type: http
      scheme: bearer
    cookieAuth:
      type: apiKey
      in: cookie
      name: token  

paths:
  /instagram:
    get:
      tags:
        - Object
      summary: returns instagram profile informations
      responses:
        '200': #status code
         description: success
         content:
          application/json:
            schema:
              type: object
              properties:
                username:
                  type: string
                followers:
                  type: string
                following:
                  type: string
        '400':
          description: Bad request
        '500':
          description: 'Internal server error'            
    post:
      summary: returns instagram profile informations
      responses:
        '200': #status code
         description: success
         content:
          application/json:
            schema:
              type: object
              properties:
                username:
                  type: string
                followers:
                  type: string
                following:
                  type: string

  /{token}:
    get:
      tags:
        - String
      summary: returns whatever parameter we pass
      parameters:
        - name: token
          in: path
          default: 10
          schema:
            type: string
      responses:
        '200': #status code
         description: success
         content:
          application/json:
            schema:
              type: string
              example: "my string"
        '400': #status code
         description: Bad request
        '500': #status code
         description: Internal server error

  /courses:
    get:
      tags:
        - Array
      summary: Returns list of all courses
      responses:
        '200':
          description: success
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    id:
                      type: number
                    name:
                      type: string
                    price:
                      type: string    
        '400':
          description: Bad request
        '500':
          description: Internal server error
    post:
      tags:
        - Array
      summary: Add a new course
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                  id:
                    type: number
                  name:
                    type: string
                  price:
                    type: string  
      responses:
        '200':
          description: created successfully
          content:
            application/json:
              schema:
                  type: boolean
        '400':
          description: Bad request
        '500':
          description: Internal server error

  /courseObject:
    get:
      tags:
        - Object
      summary: Returns list of all courses
      responses:
        '200':
          description: success
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: number
                  name:
                    type: string
                  price:
                    type: string    
        '400':
          description: Bad request
        '500':
          description: Internal server error    

  /mycourse/{courseId}:
    get:
      tags:
        - Object
      summary: Returns course using parameter courseId
      parameters:
        - name: courseId
          in: path
          required: true
          default: 22
          schema:
            type: string      
      responses:
        '200':
          description: success
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: number
                  name:
                    type: string
                  price:
                    type: string    
        '400':
          description: Bad request
        '500':
          description: Internal server error

  /coursesquery:
    get:
      tags:
        - Object
      summary: retrieve the query object
      parameters:
        - name: location
          in: query
          required: true
          schema:
            type: string
            enum: [chennai, erode, cbe]
        - name: device
          in: query
          required: true
          schema:
            type: string
            enum: [web, mobile]
      responses:
        '200':
          description: success
          content:
            application/json:
              schema:
                type: object
                properties:
                  location:
                    type: string
                  device:
                    type: string  
        '400':
          description: Bad request
        '500':
          description: Internal server error 

  /courseupload:
    post:
      tags:
        - string
      summary: Uploading course image
      parameters:
        - in: header
          name: auth
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string  
                  format: binary
      responses:
        '200':
          description: success
        '400':
          description: Bad request
        '500':
          description: Internal server error 